modularize cycle files and fixed scrollview
