+++ src/scenes/cycle/emotionData.js
@@ -1,14 +1,14 @@
-const data = [
+const emotionData = [
     {
-      title: "What is the Menstrual Cycle?",
-      description: "Some Article",
-      source: 'https://2rdnmg1qbg403gumla1v9i2h-wpengine.netdna-ssl.com/wp-content/uploads/sites/3/2018/11/MenstrualCycleAges-1176498451x533-1-650x428.jpg',
-      link: 'https://helloclue.com/articles/cycle-a-z/the-menstrual-cycle-more-than-just-the-period'
-    },
+        title: "Stress and the Menstrual Cycle",
+        description: "Some Article",
+        source: 'https://www.bannerhealth.com/healthcareblog/-/media/images/project/healthcareblog/hero-images/2022/03/good-stress-bad-stress.ashx',
+        link: 'https://helloclue.com/articles/cycle-a-z/stress-your-period'
+      },
     {
       title: "Irregular Menstrual Cycles",
       source: 'https://yesmomfertility.com/blog/wp-content/uploads/2019/09/unpredictable-mentrual-cycle-yesmom.-jpg.jpg',
-      link: 'https://helloclue.com/articles/cycle-a-z/what-is-an-irregular-menstrual-cycle'
+      link: 'https://images.ctfassets.net/juauvlea4rbf/4Bt99vnD1SewYwaoOE0I80/a1710bb768c87b58095dd04a2e76f2bf/Contenful__12_.svg'
     },
     {
       title: "Relieving Period Cramps",
@@ -17,4 +17,4 @@ const data = [
       link: 'https://helloclue.com/articles/cycle-a-z/period-cramps-101-why-menstrual-cramps-and-pain-occur-and-how-to-relieve'
     },
   ];
-  export default data;
+  export default emotionData;
